+++ lib/mix/tasks/hex.user.ex
@@ -38,7 +38,7 @@ defmodule Mix.Tasks.Hex.User do
   by you and will give access to your private resources, do not share this key with anyone. For
   keys that will be shared by organization members use `mix hex.organization key` instead. By
   default this command sets the `api:write` permission which allows write access to the API,
-  it can be overriden with the `--permission` flag.
+  it can be overridden with the `--permission` flag.
 
       mix hex.user key generate
 
@@ -76,7 +76,7 @@ defmodule Mix.Tasks.Hex.User do
 
   ## Reset user account password
 
-  Starts the process for reseting account password.
+  Starts the process for resetting account password.
 
       mix hex.user reset_password account